[us-11] Move athlete router to its own structure
