+++ src/router.py
@@ -2,6 +2,7 @@ from fastapi import FastAPI, APIRouter
 
 from src.configs.environment import get_environment_variables
 from src.routers.account import account_router
+from src.routers.athlete import athlete_router
 from src.routers.team import team_router
 from src.routers.token import token_router
 
@@ -26,3 +27,4 @@ def init_routes(app: FastAPI):
     app.include_router(account_router, prefix=BASE_PATH)
     app.include_router(team_router, prefix=BASE_PATH)
     app.include_router(token_router, prefix=BASE_PATH)
+    app.include_router(athlete_router, prefix=BASE_PATH)
